disambiguation/pyproject.toml
[tool.poetry]
name = "inspire-disambiguation"
version = "0.1"
description = "Inspire author disambiguation project."
authors = ["CERN <admin@inspirehep.net>"]
license = "MIT License"


[tool.poetry.dependencies]
python = ">=3.6,<4"
attrs = '~=19.0,>=19.1.0'
beard = "0.2.2"
click = "*"
httplib2 = '~=0.0,>=0.12.0'
langdetect = '~=1.0,>=1.0.7'
redis = '~=3.0'
urllib3 = '==1.23'
elasticsearch_dsl = ">=7.0.0,<8.0.0"
inspire-utils = "^3.0"
inspire-schemas = "^61.0"
deepdiff = "^4.0"
structlog = "^19.1"
structlog-sentry = "^1.1"
requests = "^2.22"


[tool.poetry.dev-dependencies]
flake8-future-import = '~=0.0,>=0.4.4'
coverage = ">=4.4.1"
isort = ">=4.3"
mock = ">=2.0.0"
pydocstyle = ">=2.0.0"
pytest = ">=3.3.1,<4.0.0"
pytest-cov = "==2.6.1"
pytest-mock = ">=1.6.0"
pytest-random-order = ">=0.5.4"
pytest-datadir = "*"
pytest-runner = ">=3.0.0,<5"
black = "==18.9b0"
requests-mock = "~=1.5.2,>=1.5.2"
pylint = "*"
flake8 = "*"
virtualenv = "<16.5.0"
ipdb = "^0.12.0"

[build-system]
requires = ["poetry==0.12.16"]
build-backend = "poetry.masonry.api"
